Add files via upload
Fix Script that will auto create fields in a given table.  Useful when creating several fields on a table with lots of records and it takes excessively long to create all those fields.  Just fill out the JSON at the bottom and let it fly.  Now you can do something else instead of waiting for each field to be created to start the next one.